# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/methods/get": {
      "get": {
        "tags": [
          "Methods"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a 200 OK response",
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/methods/post": {
      "post": {
        "tags": [
          "Methods"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a 200 OK response",
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/methods/delete": {
      "delete": {
        "tags": [
          "Methods"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a 200 OK response",
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/methods/head": {
      "head": {
        "tags": [
          "Methods"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a 200 OK response",
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/methods/options": {
      "options": {
        "tags": [
          "Methods"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a 200 OK response",
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/methods/patch": {
      "patch": {
        "tags": [
          "Methods"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a 200 OK response",
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/methods/put": {
      "put": {
        "tags": [
          "Methods"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a 200 OK response",
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/auth/hiddenbasicAuth/{user}/{pass}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "GET request with Basic Auth",
        "description": "Testing basic header authentication with challenge enabled. (Keep asking until you enter correct)",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pass",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "401": {
            "description": "UnAuth Request."
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/auth/basicAuth/{user}/{pass}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "GET request with Basic Auth",
        "description": "Testing basic header authentication with challenge enabled. (Keep asking until you enter correct)",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pass",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "401": {
            "description": "UnAuth Request."
          }
        }
      }
    },
    "/auth/digestAuth/{user}/{pass}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "GET request with Digest Auth",
        "description": "Testing digest header authentication with challenge enabled. (Keep asking until you enter correct)",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pass",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "401": {
            "description": "UnAuth Request."
          }
        }
      }
    },
    "/custom/get/{responseCode}": {
      "get": {
        "tags": [
          "Custom"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a responseCode",
        "parameters": [
          {
            "name": "responseCode",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/custom/post/{responseCode}": {
      "post": {
        "tags": [
          "Custom"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a responseCode",
        "parameters": [
          {
            "name": "responseCode",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          },
          "100-199": {
            "description": "Informational responses"
          },
          "200-299": {
            "description": "Successful responses"
          },
          "300 - 399": {
            "description": "Redirection messages"
          },
          "400-499": {
            "description": "Client error responses"
          },
          "500-599": {
            "description": "Server error responses "
          }
        }
      }
    },
    "/custom/delete/{responseCode}": {
      "delete": {
        "tags": [
          "Custom"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a responseCode",
        "parameters": [
          {
            "name": "responseCode",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/custom/head/{responseCode}": {
      "head": {
        "tags": [
          "Custom"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a responseCode",
        "parameters": [
          {
            "name": "responseCode",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/custom/options/{responseCode}": {
      "options": {
        "tags": [
          "Custom"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a responseCode",
        "parameters": [
          {
            "name": "responseCode",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/custom/patch/{responseCode}": {
      "patch": {
        "tags": [
          "Custom"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a responseCode",
        "parameters": [
          {
            "name": "responseCode",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/custom/put/{responseCode}": {
      "put": {
        "tags": [
          "Custom"
        ],
        "summary": "Returns Request Headers",
        "description": "Returns Request Headers as a responseCode",
        "parameters": [
          {
            "name": "responseCode",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Auth Successful."
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  }
}